+++ src/components/Card/style.tsx
@@ -18,10 +18,6 @@ const Style = styled.div`
 
   &.card {
     box-shadow: 0 1px 3px rgba(0,0,0,0.12), 0 1px 2px rgba(0,0,0,0.24);
-
-    :active {
-      background: #9FFFB0;
-    }
   
     :hover {
       box-shadow: 0 14px 28px rgba(0,0,0,0.25), 0 10px 10px rgba(0,0,0,0.22);